requirements.txt
fastapi>=0.111.0
uvicorn==0.29.0
requests==2.31.0
pydantic>=2.7.4
pydantic-settings>=2.4.0
langchain-core>=0.3.69
langchain-community==0.3.27
langchain-huggingface==0.3.0
chromadb==0.4.24
python-multipart==0.0.9
PyMuPDF==1.23.22
faiss-cpu==1.7.4
sentence-transformers>=2.7.0
numpy<2.0.0
transformers==4.41.2
torch>=2.0.0
ollama